Refactor JsonAdaptedModuleTest & JsonUtilTest
