+++ src/test/java/seedu/tarence/commons/util/JsonUtilTest.java
@@ -7,6 +7,8 @@ import java.nio.file.Path;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.tarence.model.util.SampleDataUtil;
+import seedu.tarence.storage.JsonAdaptedModule;
 import seedu.tarence.testutil.SerializableTestClass;
 import seedu.tarence.testutil.TestUtil;
 
@@ -14,6 +16,21 @@ import seedu.tarence.testutil.TestUtil;
  * Tests JSON Read and Write
  */
 public class JsonUtilTest {
+
+    public static final String INVALID_STUDENT_STRING_MISSING_STUDENT_NAME = "[studentEmail=e0035152@u.nus.edu.sg, "
+            + "studentMatricNumber=OptionalA0155413M, studentNusnetId=OptionalE0031550, "
+            + "studentModuleCode=CS1010S, studentTutorialName=Lab Session}]";
+    public static final String INVALID_STUDENT_STRING_WRONG_ORDER_OF_FIELDS = "[studentName=Alice, "
+            + "studentEmail=e0035152@u.nus.edu.sg, "
+            + "studentMatricNumber=OptionalA0155413M, studentNusnetId=OptionalE0031550, "
+            + "studentTutorialName=Lab Session, studentModuleCode=CS1010S}]";
+    public static final String INVALID_TUTORIAL_STRING_MISSING_TUTORIAL_NAME = "{tutorialDayOfWeek=MONDAY, "
+            + "studentListString=[], tutorialModuleCode=CS1010S, tutorialStartTime=12:00, tutorialDuration=PT2H, "
+            + "tutorialWeeks=[1, 4, 7]}";
+    public static final String INVALID_TUTORIAL_STRING_WRONG_ORDER_OF_FIELDS = "{tutorialName=Sectional, "
+            + "tutorialDayOfWeek=MONDAY, "
+            + "studentListString=[], tutorialModuleCode=CS1010S, tutorialStartTime=12:00, tutorialDuration=PT2H, "
+            + "tutorialWeeks=[1, 4, 7]}";
 
     private static final Path SERIALIZATION_FILE = TestUtil.getFilePathInSandboxFolder("serialize.json");
 
@@ -37,6 +54,34 @@ public class JsonUtilTest {
         assertEquals(serializableTestClass.getName(), SerializableTestClass.getNameTestValue());
         assertEquals(serializableTestClass.getListOfLocalDateTimes(), SerializableTestClass.getListTestValues());
         assertEquals(serializableTestClass.getMapOfIntegerToString(), SerializableTestClass.getHashMapTestValues());
+    }
+
+    @Test
+    public void isValidStudentString_studentStringWithoutNameField_returnsFalse() {
+        JsonAdaptedModule module = new JsonAdaptedModule(SampleDataUtil.getSampleModule());
+        assertEquals(false, seedu.tarence.commons.util.JsonUtil
+                .isValidStudentString(INVALID_STUDENT_STRING_MISSING_STUDENT_NAME));
+    }
+
+    @Test
+    public void isValidStudentString_studentStringInWrongOrder_returnsFalse() {
+        JsonAdaptedModule module = new JsonAdaptedModule(SampleDataUtil.getSampleModule());
+        assertEquals(false, seedu.tarence.commons.util.JsonUtil
+                .isValidStudentString(INVALID_STUDENT_STRING_WRONG_ORDER_OF_FIELDS));
+    }
+
+    @Test
+    public void isValidTutorialString_tutorialStringInWrongOrder_returnsFalse() {
+        JsonAdaptedModule module = new JsonAdaptedModule(SampleDataUtil.getSampleModule());
+        assertEquals(false, seedu.tarence.commons.util.JsonUtil
+                .isValidTutorialString(INVALID_TUTORIAL_STRING_WRONG_ORDER_OF_FIELDS));
+    }
+
+    @Test
+    public void isValidTutorialString_tutorialStringWithoutNameField_returnsFalse() {
+        JsonAdaptedModule module = new JsonAdaptedModule(SampleDataUtil.getSampleModule());
+        assertEquals(false, seedu.tarence.commons.util.JsonUtil
+                .isValidTutorialString(INVALID_TUTORIAL_STRING_MISSING_TUTORIAL_NAME));
     }
 
     //TODO: @Test jsonUtil_readJsonStringToObjectInstance_correctObject()
